requirements.txt
# Note that Pinax was used by the original WikiTrans website, but has not yet been integrated. Several libraries have yet to be integrated.
--find-links=http://pypi.pinaxproject.com
--find-links=http://pypi2.pinaxproject.com

python-dateutil>=1.4.1
python-openid>=2.2.4
python-yadis>=1.1.0
simplejson>=2.0.9
pytz>=2009l

django-notification>=0.1.4
django-openid>=0.2.0
django-email-confirmation>=0.1.3
django-mailer>=0.1.0
django-announcements>=0.1.0
django-pagination>=1.0.5.1
django-timezones>=0.1.4
django-ajax-validation>=0.1.3
django-uni-form>=0.6.0
template_utils>=0.4p2
django-staticfiles>=0.1.2
django-avatar>=1.0.3
django-gravatar>=0.1.0
django-piston>=0.2.2

boto>=1.9b
pyango_view>=0.1
wikipydia>=0.1.4
goopytrans>=0.1.1
pyyaml>=3.09
nltk==2.0b7
werkzeug>=0.5.1
polib==0.5.3
mwlib>=0.8
cython>=0.11
BeautifulSoup>=3.0
apyrtium>=0.1
python-Levenshtein>=0.10

# ISO-639 conversion
pycountry>=0.12.1

# XML-RPC Integration with MT Server Land
httplib2>=0.6.0
